+++ app/components/logo/index.test.js
@@ -1,12 +1,17 @@
 import {mount} from 'enzyme'
 import React from 'react'
-import 'ignore-styles'
 
 import Logo from './index'
+import {Icon} from 'progressive-web-sdk/dist/components/icon'
 
 /* eslint-disable newline-per-chained-call */
 
 test('Logo renders without errors', () => {
     const wrapper = mount(<Logo />)
+    expect(wrapper.length).toBe(1)
+})
+
+test('Icon renders without errors', () => {
+    const wrapper = mount(<Icon />)
     expect(wrapper.length).toBe(1)
 })