+++ tests/emptyStringMock.js
@@ -1,0 +1,4 @@
+// This is required for successfully importing SVG assets in the Jest testing
+// environment.
+
+module.exports = '' // eslint-disable-line import/no-commonjs